domene/src/main/java/no/nav/foreldrepenger/selvbetjening/util/DateUtil.java
package no.nav.foreldrepenger.selvbetjening.util;

import static java.time.LocalDate.now;
import static java.time.Month.OCTOBER;
import static java.time.temporal.TemporalAdjusters.firstDayOfYear;

import java.time.LocalDate;

public final class DateUtil {

    private DateUtil() {
    }

    public static boolean erNyopprettet(LocalDate fom) {
        return erNyopprettet(LocalDate.now(), fom);
    }

    static boolean erNyopprettet(LocalDate nå, LocalDate fom) {
        return fom.isAfter(now().minusYears(nå.isAfter(LocalDate.of(nå.getYear(), OCTOBER, 20)) ? 3 : 4)
                .with(firstDayOfYear()).minusDays(1));
    }

}
